Add routine to get upcoming pick ups
Wrote stored procedure for getting the user emails for approved
requests that are starting soon. It takes an input for the number of
days to look forward to. The test data has an approved request coming
up on Dec. 12 (which is 50 days from today). So if you call this
routine with an input of 50 you will get that users email. This
procedure could be used by some periodic process to look for upcoming
pick ups that are 7 days away then email those users to remind them.